Tests, Finished our Int Pushing, and Began our Var Checking
Tried some of the tests out, not working, probably because I haven't finished the main code. Int pushing looks good, and implemented a multiplication and divide for PEMDAS, also started our variable checker, although I think I want to check for exceptions rather than trying to check if it's true or not.